Nuke the old structurizer
