+++ crates/rustc_codegen_spirv/src/link.rs
@@ -539,7 +539,6 @@ fn do_link(
         inline: legalize,
         mem2reg: legalize,
         structurize: env::var("NO_STRUCTURIZE").is_err(),
-        use_new_structurizer: env::var("OLD_STRUCTURIZER").is_err(),
         emit_multiple_modules,
     };
